requirements.txt
google-generativeai
python-pptx
python-dotenv
comtypes
streamlit
requests
moviepy
Pillow
edge-tts